Added implementation of jsonl, extended cache, helper test script
jsonl is implemented to read files which have a single valid json
object on each line.

Sequence.cache() was extended so that after evaluating the lineage,
the lineage can be deleted/forgotten. This is helpful when using
functional sequences is convenient within the package, but outside
users should not see the effects of internal transformations

Added a test runner script to facilitate pre-commit/push testing