Fixed the performance regression in LazyFile from #45
The performance regression was caused by implementing the next() or
__next__() methods which were being called instead of the built in C
implementation from the object file. The fix is to only implement
__iter__() to return the file since it is its own iterator.
Additionally this is wrapped with a keyword with block and returns
lines with yield to make sure that after the iteration the file is
closed. This also makes it so that calls __iter__ are completely
independent where before they were related.

Lastly, LazyFile was renamed to ReusableFile since the name more clearly
communicates its purpose.