Added to_pandas support
Support added for usefulness in some recent work. Functionality
is similar in spirit/implementation as Spark's DataFrame.toPandas

Requirement for pandas is not added, only on the to_pandas call is
the import attempted. Added pandas as a Travis install, if this takes
too long to build/test, will look into pip cache